move shell utilities to cli folder
