+++ src/cli/kubernetes/kubernetesTools.ts
@@ -4,8 +4,8 @@ import { Uri, window } from 'vscode';
 import * as kubernetes from 'vscode-kubernetes-tools-api';
 import { AzureConstants } from '../azure/azureTools';
 import { globalState, setVSCodeContext, telemetry } from '../../extension';
-import { output } from '../../shell/output';
-import { shellCodeError } from '../../shell/shell';
+import { output } from '../shell/output';
+import { shellCodeError } from '../shell/shell';
 import { TelemetryErrorEventNames } from '../../types/telemetryEventNames';
 import { Errorable, failed, succeeded } from '../../types/errorable';
 import { ContextId } from '../../types/extensionIds';
