+++ src/cli/shell/terminal.ts
@@ -1,5 +1,5 @@
 import { Disposable, ExtensionContext, Terminal, window } from 'vscode';
-import { extensionContext } from '../extension';
+import { extensionContext } from '../../extension';
 
 const terminalName = 'gitops';
