+++ src/commands/commands.ts
@@ -1,6 +1,6 @@
 import { commands, Disposable, ExtensionContext, Uri, window } from 'vscode';
 import { telemetry } from '../extension';
-import { showOutputChannel } from '../shell/output';
+import { showOutputChannel } from '../cli/shell/output';
 import { TelemetryErrorEventNames } from '../types/telemetryEventNames';
 import { refreshAllTreeViews, refreshResourcesTreeViews } from '../ui/treeviews/treeViews';
 import { CommandId } from '../types/extensionIds';
