+++ src/commands/installFluxCli.ts
@@ -8,9 +8,9 @@ import { commands, window } from 'vscode';
 import { Errorable, failed, succeeded } from '../types/errorable';
 import { globalState } from '../extension';
 import { GlobalStateKey } from '../data/globalState';
-import { output } from '../shell/output';
-import { Platform, shell, shellCodeError } from '../shell/shell';
-import { runTerminalCommand } from '../shell/terminal';
+import { output } from '../cli/shell/output';
+import { Platform, shell, shellCodeError } from '../cli/shell/shell';
+import { runTerminalCommand } from '../cli/shell/terminal';
 import { appendToPathEnvironmentVariableWindows, createDir, deleteFile, downloadFile, getAppdataPath, moveFile, readFile, unzipFile } from '../utils/fsUtils';
 import { refreshAllTreeViews } from '../ui/treeviews/treeViews';
