+++ src/commands/pullGitRepository.ts
@@ -5,7 +5,7 @@ import safesh from 'shell-escape-tag';
 import { commands, Uri, window } from 'vscode';
 import { telemetry } from '../extension';
 import { checkGitVersion } from '../cli/checkVersions';
-import { shell } from '../shell/shell';
+import { shell } from '../cli/shell/shell';
 import { TelemetryErrorEventNames } from '../types/telemetryEventNames';
 import { quoteFsPath } from '../utils/fsUtils';
 import { GitRepositoryNode } from '../ui/treeviews/nodes/gitRepositoryNode';
